Move TimeTickChannel out of utils.
