+++ infrastructure/time/build.gradle
@@ -1,3 +1,4 @@
 dependencies {
+  implementation project(':infrastructure:events')
   testFixturesImplementation project(':infrastructure:unsigned')
 }
