+++ services/timer/src/main/java/tech/pegasys/teku/services/timer/TimerService.java
@@ -26,9 +26,9 @@ import org.quartz.SchedulerFactory;
 import org.quartz.SimpleTrigger;
 import org.quartz.impl.StdSchedulerFactory;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.time.TimeTickChannel;
 import tech.pegasys.teku.service.serviceutils.Service;
 import tech.pegasys.teku.service.serviceutils.ServiceConfig;
-import tech.pegasys.teku.util.time.channels.TimeTickChannel;
 
 public class TimerService extends Service {
